Supervisor working

+++ supervisor.js
@@ -4,8 +4,10 @@ const path = require("path");
 
 const projectDir = path.resolve(__dirname);
 const srcDir = path.join(projectDir, "src");
+const triggerFile = path.join(projectDir, "deploy.trigger");
 const entryFile = path.join(srcDir, "index.js");
 let currentChild = null;
+let isDeploying = false;
 let firstRun = true;
 
 async function runCmd(cmd, args) {
@@ -16,24 +18,30 @@ async function runCmd(cmd, args) {
 }
 
 async function deploy({ checkGit = false, checkInstall = false }) {
-	console.log("===> Deploying fluxloader server");
+	if (isDeploying) {
+		console.log("Deploy already in progress, skipping deploy");
+		return;
+	}
+	isDeploying = true;
+
+	console.log("==> Deploying fluxloader server");
 
 	if (currentChild) {
-		console.log("=> Killing old server process");
+		console.log("> Killing old server process");
 		currentChild.kill("SIGTERM");
 	}
 
 	if (checkGit) {
-		console.log("=> Checking for git updates");
+		console.log("> Checking for git updates");
 		await runCmd("git", ["pull"]);
 	}
 
 	if (checkInstall) {
-		console.log("=> Checking for npm install");
+		console.log("> Checking for npm install");
 		await runCmd("npm", ["i"]);
 	}
 
-	console.log("=> Starting new server process");
+	console.log("> Starting new server process");
 	const child = spawn(process.execPath, [entryFile], { cwd: srcDir, stdio: "inherit" });
 
 	child.on("exit", (code, signal) => {
@@ -42,15 +50,29 @@ async function deploy({ checkGit = false, checkInstall = false }) {
 	});
 
 	currentChild = child;
+	isDeploying = false;
 }
 
 (async () => {
-	fs.watchFile(path.join(projectDir, "deploy.trigger"), async () => {
+	if (fs.existsSync(triggerFile)) {
+		try {
+			fs.unlinkSync(triggerFile);
+		} catch (e) {
+			console.error("Failed to unlink deploy.trigger:", e);
+			return;
+		}
+	}
+
+	fs.watchFile(triggerFile, async () => {
+		if (isDeploying) {
+			console.log("Deploy already in progress, ignoring trigger");
+			return;
+		}
 		if (firstRun) {
 			await deploy({ checkGit: false, checkInstall: false }).catch((e) => console.error(e));
 			firstRun = false;
 		} else {
-			console.log("<=== Detected deploy.trigger change! redeploying...\n");
+			console.log("<== Detected deploy.trigger change! redeploying...\n");
 			deploy({ checkGit: true, checkInstall: true }).catch((e) => console.error(e));
 		}
 	});
